Moved NumericLiteralType
If it's used in exactly one method, which package should it be in?
